fix : pull하기 전 변경사항
